Increase int size for tweets

+++ src/bitcoin_acks/models/tweets.py
@@ -1,6 +1,6 @@
 import datetime
 
-from sqlalchemy import Column, DateTime, Integer
+from sqlalchemy import BIGINT, Column, DateTime
 
 from bitcoin_acks.database.base import Base
 
@@ -8,7 +8,7 @@ from bitcoin_acks.database.base import Base
 class Tweets(Base):
     __tablename__ = 'tweets'
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BIGINT, primary_key=True)
     timestamp = Column(DateTime, nullable=False,
                        default=datetime.datetime.utcnow, )
-    pull_request_id = Column(Integer, nullable=False, unique=True)
+    pull_request_id = Column(BIGINT, nullable=False, unique=True)
